// Admin/src/models/Repair/RepairType.Model.ts
export interface IRepairTypeModel {
    RepairTypeName:string
    OrderId:number
    Active: boolean
    AllocatedPrice: number
    Created: any
    CreatedBy: number
    DisplayInList: boolean
    DisplayName: string
    Enabled: any
    EnumName: string
    Id: number
    IsValid: boolean
    Modified: any
    ModifiedBy: number
    Name: string
    OfferPrice: number
    RowOrder: number
    ThumbnailPath: string
    ValidationErrors: {}
}